% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_charts.R
\name{gender_profile}
\alias{gender_profile}
\title{line chart to show the number of headcount by financial year by gender.}
\usage{
gender_profile(x)
}
\arguments{
\item{x}{Input headcount_data S3 class object.}
}
\value{
Returns a highchart or htmlwidget object.
}
\description{
\code{headcount_data} is the S3 class used for gender related 
summary of workforce
}
\examples{

workforce <- nhsbsaGPG::headcount_data(nhsbsaGPG::headcount)
nhsbsaGPG::gender_profile(workforce)

}
