% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-charts.R
\name{gpg_bar}
\alias{gpg_bar}
\title{Highcharter bar chart. This chart will show columns horizontally.}
\usage{
gpg_bar(x, xvar, yvar, yaxis_title)
}
\arguments{
\item{x}{Input df_hrrate_afc or df_hrrate_dir data frame.}

\item{xvar}{e.g. mean_paygap, median_paygap}

\item{yvar}{AfC band, directorate}

\item{yaxis_title}{Y axis title}
}
\value{
Returns a highchart or htmlwidget object.
}
\description{
{df_hrrate_afc} or {df_hrrate_dir} data frame is used for bar chart.
}
