% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-charts.R
\name{gpg_column}
\alias{gpg_column}
\title{Highcharter column chart. This chart will show columns vertically}
\usage{
gpg_column(x, xvar = "period", yvar, yaxis_title, export_filename)
}
\arguments{
\item{x}{Input paygap data frame.}

\item{xvar}{"period" default}

\item{yvar}{e.g. mean_paygap, median_paygap}

\item{yaxis_title}{Y axis title}
}
\value{
Returns a highchart or htmlwidget object.
}
\description{
{paygap} data frame is used for bar chart.
}
