% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpg_data_class.R
\name{gpg_data}
\alias{gpg_data}
\title{S3 gender pay gap class}
\usage{
gpg_data(x, log_level = futile.logger::WARN, eda = FALSE)
}
\arguments{
\item{x}{Input data frame.}

\item{log_level}{keep it WARN}

\item{eda}{If TRUE base R plot shows in the Viewer}
}
\value{
If the class is not instantiated correctly, nothing is returned.
}
\description{
\code{gpg_data} is the class used for the creation of
headcount, hourly rate by further breakdown.
}
\details{
The \code{gpg_data} class expects a \code{data.frame} with at
least seven columns: period, gender, hourly_rate, quartile, afc_band,
directorate.

Once initiated, the class has six slots:
\code{df_hdcnt}: data frame contains overall headcount
\code{df_hdcnt_gender}: data frame contains headcount
\code{df_hdcnt_afc}: data frame contains headcount by AFC band
\code{df_hdcnt_dir}: data frame contains headcount by directorate
\code{df_hrrate}: data frame contains mean, median hourly rate
\code{df_hrrate_afc}: data frame contains mean, median hourly rate by AFC band
\code{df_hrrate_dir} data frame contains mean, median hourly rate by directorate
\code{ending_fy}: a character vector containing ending reporting period
(e.g. 31 March 2023). This uses for introduction paragraph
}
\examples{
\dontrun{

library(nhsbsaGPG)
df <- gpg_data(afc_staff)
}

}
