% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-charts.R
\name{gpg_dumbbell}
\alias{gpg_dumbbell}
\title{Highcharter dumbbell chart.}
\usage{
gpg_dumbbell(x, low, high, xaxis_category, yaxis_title, export_filename)
}
\arguments{
\item{x}{Input {df_hrrate_afc} or {df_hrrate_dir} data frame.}

\item{yaxis_title}{Y axis title}

\item{xvar}{"period" default}

\item{yvar}{e.g. mean_paygap, median_paygap}
}
\value{
Returns a highchart or htmlwidget object.
}
\description{
df_hrrate_afc or df_hrrate_dir data frame.
}
