% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_charts.R
\name{gpg_pyramid}
\alias{gpg_pyramid}
\title{Highcharter bar chart to create pyramid chart. This chart
shows gender pay gap information by AFC band in NHSBSA}
\usage{
gpg_pyramid(x, xvar = "afc_band", yvar, yaxis_title)
}
\arguments{
\item{x}{Input data frame from \code{gpg_data} S3 class object.}

\item{xvar}{"afc_band" default value}

\item{yvar}{headcount/mean hourly/median hourly pay}

\item{yaxis_title}{Y axis title}
}
\value{
Returns a highchart or htmlwidget object.
}
\description{
\code{gpg_data} is the S3 class used for trend
}
