% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-charts.R
\name{gpg_stack}
\alias{gpg_stack}
\title{Highcharter column chart to create stacked column chart. This chart
shows proportion of males and females in each quartile pay band.}
\usage{
gpg_stack(x, xvar, yvar, groupvar, yaxis_title, export_filename)
}
\arguments{
\item{x}{Input quartile data frame.}

\item{xvar}{"afc_band" default value}

\item{yvar}{headcount/mean hourly/median hourly pay}

\item{groupvar}{group by variable}

\item{yaxis_title}{Y axis title}
}
\value{
Returns a highchart or htmlwidget object.
}
\description{
{quartile} data frame is used for stacked column chart.
}
