% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-charts.R
\name{gpg_trend}
\alias{gpg_trend}
\title{Highcharter line chart to show the number of headcount by financial
year by gender.}
\usage{
gpg_trend(
  x,
  xvar = "period",
  yvars,
  series_names,
  yaxis_title,
  yaxis_label,
  colpalette
)
}
\arguments{
\item{x}{Input data frame from \code{gpg_data} S3 class object.}

\item{xvar}{"period", default}

\item{yvars}{data frame converts to list and each list element to create line}

\item{series_names}{If user wants to give different series name for highchart legend}

\item{yaxis_title}{Y axis title}

\item{yaxis_label}{Indication of percentage or number}

\item{colpalette}{custom palette}
}
\value{
Returns a highchart or htmlwidget object.
}
\description{
\code{gpg_data} is the S3 class used for trend
}
