% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{headcount}
\alias{headcount}
\title{NHSBSA employee headcount}
\format{

}
\usage{
headcount
}
\description{
A dataset containing NHSBSA employee headcount
since financial year 2017/18, split by gender, AFC band 
and FTE (full time or part time)
}
\keyword{datasets}
