% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quartile}
\alias{quartile}
\title{NHSBSA employee quartile}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50 rows and 5 columns.
}
\usage{
quartile
}
\description{
A dataset containing NHSBSA employee hourly pay
by reporting period from 2018 (eg. snapshot 31/03/2018 for 2018/19 report etc)
by quartiles
}
\details{
\itemize{
  \item period. 2018/19, 2019/20 etc character
  \item quartile. split hourly pay into quartiles
  \item female. number of female employees in each quartile
  \item male. number of male employees in each quartile
  \item quartile. split hourly_rate by quartile by gender
 }

 @docType data
 @keywords datasets
 @name quartile
 @usage quartile
 @format data frame with employee gender hourly pay by quartiles
}
\keyword{datasets}
