% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afc_staff}
\alias{afc_staff}
\title{NHSBSA employee afc_staff}
\format{
A data frame with gender pay gap information
}
\usage{
afc_staff
}
\description{
A dataset containing NHSBSA employee hourly pay
by reporting period split by gender, AfC band, directorate.
Input data frame for gpg_class
\itemize{
  \item period. 2018/19, 2019/20 etc character
  \item gender. Male or Female, character
  \item headcount. employee headcount used for aggregation
  \item hourly_rate. hourly rate as shown pay slip
  \item quartile. split hourly_rate by quartile by gender
  \item afc_band. AFC band
  \item directorate. NHSBSA directorate
}
}
\keyword{datasets}
